[MP][Coordinator] Move the eviction controller's endpoints onto it
